Changed ConstantTrendAnomalyDetector to ModeAnomalyDetector, using new approach
